+++ packages/app/src/cli/models/extensions/ui-specifications/ui_extension.test.ts
@@ -1,6 +1,6 @@
-import {ExtensionInstance, ExtensionSpec, specForType} from '../extensions.js'
 import {configurationFileNames} from '../../../constants.js'
 import * as loadLocales from '../../../utilities/extensions/locales-configuration.js'
+import {UIExtensionInstance, UIExtensionSpec, uiSpecForType} from '../ui.js'
 import {describe, expect, test, vi} from 'vitest'
 import {file, path} from '@shopify/cli-kit'
 import {err, ok} from '@shopify/cli-kit/node/result'
@@ -13,7 +13,7 @@ describe('ui_extension', async () => {
 
   async function getTestUIExtension({directory, extensionPoints}: GetUIExtensionProps) {
     const configurationPath = path.join(directory, configurationFileNames.extension.ui)
-    const specification = (await specForType('ui_extension')) as ExtensionSpec
+    const specification = (await uiSpecForType('ui_extension')) as UIExtensionSpec
     const configuration = {
       extensionPoints,
       name: 'UI Extension',
@@ -22,11 +22,12 @@ describe('ui_extension', async () => {
       capabilities: {
         block_progress: false,
         network_access: false,
+        api_access: false,
       },
       settings: {},
     }
 
-    return new ExtensionInstance({
+    return new UIExtensionInstance({
       configuration,
       directory,
       specification,