+++ packages/app/src/cli/models/extensions/ui.ts
@@ -1,17 +1,14 @@
-import {BaseExtensionSchema, ZodSchemaType} from './schemas.js'
-import {allExtensionSpecifications} from './specifications.js'
-import {ExtensionCategory, GenericSpecification, ThemeExtension, UIExtension} from '../app/extensions.js'
+import {BaseConfigContents, ZodSchemaType} from './schemas.js'
+import {allUISpecifications} from './specifications.js'
+import {ExtensionCategory, GenericSpecification, UIExtension} from '../app/extensions.js'
 import {blocks, defualtExtensionFlavors} from '../../constants.js'
-import {id, path, schema, api, output, environment, string} from '@shopify/cli-kit'
+import {id, path, api, output, environment, string} from '@shopify/cli-kit'
 import {ok, Result} from '@shopify/cli-kit/node/result'
-
-// Base config type that all config schemas must extend.
-export type BaseConfigContents = schema.define.infer<typeof BaseExtensionSchema>
 
 /**
  * Extension specification with all the needed properties and methods to load an extension.
  */
-export interface ExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>
+export interface UIExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>
   extends GenericSpecification {
   identifier: string
   externalIdentifier: string
@@ -56,8 +53,8 @@ export interface ExtensionSpec<TConfiguration extends BaseConfigContents = BaseC
  *
  * This class holds the public interface to interact with extensions
  */
-export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseConfigContents>
-  implements UIExtension<TConfiguration>, ThemeExtension<TConfiguration>
+export class UIExtensionInstance<TConfiguration extends BaseConfigContents = BaseConfigContents>
+  implements UIExtension<TConfiguration>
 {
   entrySourceFilePath: string
   outputBundlePath: string
@@ -68,7 +65,7 @@ export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseC
   configuration: TConfiguration
   configurationPath: string
 
-  private specification: ExtensionSpec
+  private specification: UIExtensionSpec
   private remoteSpecification?: api.graphql.RemoteSpecification
 
   get graphQLType() {
@@ -108,7 +105,7 @@ export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseC
     configurationPath: string
     entryPath: string
     directory: string
-    specification: ExtensionSpec
+    specification: UIExtensionSpec
     remoteSpecification?: api.graphql.RemoteSpecification
   }) {
     this.configuration = options.configuration
@@ -176,8 +173,8 @@ export class ExtensionInstance<TConfiguration extends BaseConfigContents = BaseC
 /**
  * Find the registered spececification for a given extension type
  */
-export async function specForType(type: string): Promise<ExtensionSpec | undefined> {
-  const allSpecs = await allExtensionSpecifications()
+export async function uiSpecForType(type: string): Promise<UIExtensionSpec | undefined> {
+  const allSpecs = await allUISpecifications()
   return allSpecs.find((spec) => spec.identifier === type || spec.externalIdentifier === type)
 }
 
@@ -186,7 +183,7 @@ function remoteSpecForType(type: string): api.graphql.RemoteSpecification | unde
   return undefined
 }
 
-export function createExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>(spec: {
+export function createUIExtensionSpec<TConfiguration extends BaseConfigContents = BaseConfigContents>(spec: {
   identifier: string
   externalIdentifier: string
   partnersWebIdentifier: string
@@ -212,7 +209,7 @@ export function createExtensionSpec<TConfiguration extends BaseConfigContents = 
   ) => output.TokenizedString | undefined
   shouldFetchCartUrl?(config: TConfiguration): boolean
   hasExtensionPointTarget?(config: TConfiguration, target: string): boolean
-}): ExtensionSpec<TConfiguration> {
+}): UIExtensionSpec<TConfiguration> {
   const defaults = {
     showInCLIHelp: true,
     singleEntryPath: true,