+++ packages/app/src/cli/services/info.ts
@@ -2,7 +2,11 @@ import {outputEnv} from './app/env/show.js'
 import {AppInterface} from '../models/app/app.js'
 import {FunctionExtension, ThemeExtension, UIExtension} from '../models/app/extensions.js'
 import {configurationFileNames} from '../constants.js'
-import {allExtensionSpecifications, allFunctionSpecifications} from '../models/extensions/specifications.js'
+import {
+  allFunctionSpecifications,
+  allThemeSpecifications,
+  allUISpecifications,
+} from '../models/extensions/specifications.js'
 import {os, output, path, store, string} from '@shopify/cli-kit'
 import {checkForNewVersion} from '@shopify/cli-kit/node/node-package-manager'
 
@@ -121,11 +125,9 @@ class AppInfo {
         }
       })
     }
-    const allExtensionSpecs = await allExtensionSpecifications()
-    const allFunctionsSpecs = await allFunctionSpecifications()
-    const uiTypes = allExtensionSpecs.map((spec) => spec.identifier).filter((spec) => spec !== 'theme')
-    const themeTypes = allExtensionSpecs.map((spec) => spec.identifier).filter((spec) => spec === 'theme')
-    const functionTypes = allFunctionsSpecs.map((spec) => spec.identifier)
+    const uiTypes = (await allUISpecifications()).map((spec) => spec.identifier)
+    const themeTypes = (await allThemeSpecifications()).map((spec) => spec.identifier)
+    const functionTypes = (await allFunctionSpecifications()).map((spec) => spec.identifier)
     augmentWithExtensions(uiTypes, this.app.extensions.ui, this.uiExtensionSubSection.bind(this))
     augmentWithExtensions(themeTypes, this.app.extensions.theme, this.themeExtensionSubSection.bind(this))
     augmentWithExtensions(functionTypes, this.app.extensions.function, this.functionExtensionSubSection.bind(this))