+++ packages/cli-kit/src/testing/ui.ts
@@ -1,22 +1,24 @@
-import * as path from '../path.js'
-import {ExecaChildProcess, execaNode} from 'execa'
+import {render} from 'ink-testing-library'
 
-type Run = (fixture: string, props?: {env?: {[key: string]: unknown}}) => ExecaChildProcess<string>
+export function waitForInputsToBeReady() {
+  return new Promise((resolve) => setTimeout(resolve, 100))
+}
 
-export const run: Run = (fixture, props) => {
-  const env = {
-    ...process.env,
-    ...props?.env,
-    // we need this because ink treats the CI environment differently
-    // by only writing the last frame to stdout on unmount
-    // See more here https://github.com/vadimdemedes/ink/pull/266
-    // this way local and CI tests behave the same
-    CI: 'true',
-  }
+export function waitForChange(func: () => void, getChangingValue: () => string | number | undefined) {
+  return new Promise<void>((resolve) => {
+    const initialValue = getChangingValue()
 
-  // we want to load the compiled js directly in order avoid unnecessary transpilation
-  return execaNode(path.resolve(__dirname, `../../dist/testing/fixtures/${fixture}.js`), {
-    cwd: __dirname,
-    env,
+    func()
+
+    const interval = setInterval(() => {
+      if (getChangingValue() !== initialValue) {
+        clearInterval(interval)
+        resolve()
+      }
+    }, 10)
   })
+}
+
+export async function sendInput(renderInstance: ReturnType<typeof render>, ...inputs: string[]) {
+  await waitForChange(() => inputs.forEach((input) => renderInstance.stdin.write(input)), renderInstance.lastFrame)
 }