+++ packages/cli-hydrogen/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @shopify/cli-hydrogen
 
+## 3.28.0
+
+### Patch Changes
+
+- @shopify/cli-kit@3.28.0
+
+## 3.27.0
+
+### Patch Changes
+
+- Updated dependencies [32bbe23d]
+  - @shopify/cli-kit@3.27.0
+
 ## 3.26.0
 
 ### Patch Changes